+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- `Font::family_name` will load names with Windows Symbol encoding now.
 
 ## [0.6.1] - 2020-05-19
 ### Fixed